Update 02 Scope, Frame and Stack.rst
Removed calls to print that are not needed. For, instance if we run the following code:

```python
def foo(f):
    f = 5
    print('in foo(), f =', f)

f = 12
print(foo(f))
print('in global frame, f =', f)
```python 

the output is the following: 

in foo(), f = 5
None
in global frame, f = 12

and not the one mentioned in the document. This happens due to the following line:

```python 
print(foo(f))
```

If we remove the call to print method and leave the call to foo, then we get the output that is mentioned in the doc

in foo(), f = 5
in global frame, f = 12